update to Gene.sync_from_remotes method

+++ app/models/gene.rb
@@ -404,7 +404,7 @@ class Gene < ActiveRecord::Base
       logger.debug "[Gene.sync_with_remotes] Evaluating #{mgi_ids_to_delete.size} gene(s) for deletion..."
       mgi_ids_to_delete.each do |mgi_accession_id|
         current_gene = Gene.find_by_mgi_accession_id(mgi_accession_id)
-        if current_gene.mi_plans.size == 0
+        if current_gene.mi_plans.size == 0 && current_gene.notifications.empty?
           logger.debug "[Gene.sync_with_remotes] Deleting gene data for #{current_gene.mgi_accession_id}"
           current_gene.destroy
         end
